refactor: split main.rs into lib.rs and main.rs
This allows to integration test the rpc-extractor.
